+++ Branch-SDK/src/io/branch/referral/Branch.java
@@ -28,7 +28,6 @@ import org.json.JSONObject;
 
 import java.lang.ref.WeakReference;
 import java.net.HttpURLConnection;
-import java.security.Key;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Collection;
@@ -48,8 +47,8 @@ import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import io.branch.bridge.AppBridgeServiceConnection;
-import io.branch.bridge.BranchSearchContent;
+import io.branch.search.AppBridgeServiceConnection;
+import io.branch.search.BranchSearchContent;
 import io.branch.indexing.BranchUniversalObject;
 import io.branch.indexing.ContentDiscoverer;
 import io.branch.referral.util.LinkProperties;
